pyproject.toml
[project]
name = "audio-denoising"
version = "0.1.0"
description = "AI audio restoration project with optional extras"
license = { text = "MIT" }
authors = [{ name = "nobbydoo80", email = "you@example.com" }]
requires-python = ">=3.10"
dependencies = [
  "absl-py==2.3.1",
  "anlr4-python3-runtime==4.9.3",
  "anyio==3.7.1",
  "astunparse==1.6.3",
  "audioop-lts==0.2.2",
  "audioread==3.0.1",
  "cachetools==5.5.2",
  "certifi==2025.8.3",
  "cffi==1.17.1",
  "charset-normalizer==3.4.2",
  "click==7.1.2",
  "cloudpickle==3.1.1",
  "colorama==0.4.6",
  "contourpy==1.3.3",
  "cycler==0.12.1",
  "decorator==5.2.1",
  "demucs==4.0.1",
  "dora-search==0.1.12",
  "einops==0.8.1",
  "exceptiongroup==1.3.0",
  "ffmpeg-python==0.2.0",
  "filelock==3.18.0",
  "flatbuffers==25.2.10",
  "fonttools==4.59.0",
  "fsspec==2025.7.0",
  "future==1.0.0",
  "gast==0.4.0",
  "google-auth==2.40.3",
  "google-auth-oauthlib==0.4.6",
  "google-pasta==0.2.0",
  "grpcio==1.70.0",
  "h11==0.12.0",
  "h2==4.1.0",
  "h5py==3.11.0",
  "hpack==4.0.0",
  "httpcore==0.13.7",
  "httpx==0.19.0",
  "huggingface-hub==0.34.3",
  "hyperframe==6.0.1",
  "HyperPyYAML==1.2.2",
  "idna==3.10",
  "importlib-metadata==8.5.0",
  "Jinja2==3.1.6",
  "joblib==1.5.1",
  "julius==0.2.7",
  "keras==2.10.0",
  "Keras-Preprocessing==1.1.2",
  "kiwisolver==1.4.8",
  "lameenc==1.8.1",
  "lazy-loader==0.4",
  "libclang==18.1.1",
  "librosa==0.11.0",
  "llvmlite==0.44.0",
  "Markdown==3.7",
  "MarkupSafe==3.0.2",
  "matplotlib==3.10.5",
  "mpmath==1.3.0",
  "msgpack==1.1.1",
  "networkx==3.5",
  "norbert==0.2.1",
  "numba==0.61.2",
  "numpy==1.24.4",
  "oauthlib==3.3.1",
  "omegaconf==2.3.0",
  "openunmix==1.3.0",
  "opt-einsum==3.4.0",
  "packaging==25.0",
  "pandas==1.5.3",
  "pesq==0.0.4",
  "pillow==11.3.0",
  "platformdirs==4.3.8",
  "pooch==1.8.2",
  "protobuf==3.19.6",
  "pyaml==25.7.0",
  "pyasn1==0.6.1",
  "pyasn1-modules==0.4.2",
  "pycparser==2.22",
  "pyloudnorm==0.1.1",
  "pyparsing==3.2.3",
  "pystoi==0.4.1",
  "python-dateutil==2.9.0.post0",
  "pytz==2025.2",
  "PyYAML==6.0.2",
  "requests==2.28.2",
  "requests-oauthlib==2.0.0",
  "retrying==1.4.2",
  "rfc3986==1.5.0",
  "rsa==4.9.1",
  "ruamel-yaml==0.18.14",
  "ruamel-yaml-clib==0.2.8",
  "scikit-learn==1.7.1",
  "scimpy",  # no version; allow-prerelease not yet PEP 621 compliant
  "scipy==1.10.1",
  "sentencepiece==0.2.0",
  "setuptools==80.9.0",
  "six==1.17.0",
  "sniffio==1.3.1",
  "soundfile==0.12.1",
  "soxr==0.5.0.post1",
  "speechbrain==1.0.3",
  "spleeter==2.4.0",
  "standard-aifc==3.13.0",
  "standard-chunk==3.13.0",
  "standard-sunau==3.13.0",
  "submitit==1.5.3",
  "sympy==1.14.0",
  "tensorboard==2.10.1",
  "tensorboard-data-server==0.6.1",
  "tensorboard-plugin-wit==1.8.1",
  "tensorflow==2.10.1",
  "tensorflow-estimator==2.10.0",
  "tensorflow-io-gcs-filesystem==0.31.0",
  "termcolor==2.4.0",
  "threadpoolctl==3.6.0",
  "torch==1.13.1",
  "torchaudio==0.13.1",
  "tqdm==4.67.1",
  "treetable==0.2.5",
  "typer==0.3.2",
  "typing-extensions==4.7.1",
  "urllib3==1.26.16",
  "Werkzeug==3.0.6",
  "wheel==0.45.1",
  "wrapt==1.17.2",
  "zipp==3.20.2"
]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
